requirements.txt
# Core Framework
streamlit==1.31.0

# YouTube API
google-api-python-client==2.116.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Reddit API
praw==7.7.1

# News/Blog Scraping
newspaper3k==0.2.8
beautifulsoup4==4.12.3
lxml==5.1.0

# HTTP Requests
requests==2.31.0
urllib3==2.1.0

# Data Processing
pandas==2.2.0

# URL Validation
validators==0.22.0

# Environment Variables
python-dotenv==1.0.1

# Date/Time Handling
python-dateutil==2.8.2
